server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
psycopg2-binary==2.9.9
requests==2.31.0
python-docx==1.1.0

